[Bugfix] Fix RuntimeError: Index put requires the source and destination dtypes match

+++ vllm/model_executor/models/utils.py
@@ -428,7 +428,7 @@ def _merge_multimodal_embeddings(
             f"Attempted to assign {expr} = {flattened.shape[0]} "
             f"multimodal tokens to {num_expected_tokens} placeholders")
 
-    inputs_embeds[is_multimodal] = flattened
+    inputs_embeds[is_multimodal] = flattened.to(dtype=inputs_embeds.dtype)
     return inputs_embeds
